renamed userController to authController
